+++ activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/services/security/AuthenticationWrapper.java
@@ -1,4 +1,4 @@
-package org.activiti.cloud.services.query.security;
+package org.activiti.cloud.services.security;
 
 import org.activiti.engine.impl.identity.Authentication;
 import org.springframework.stereotype.Component;
