+++ activiti-cloud-starter-query/src/test/java/org/activiti/cloud/services/query/QueryApplication.java
@@ -30,7 +30,7 @@ import org.springframework.context.annotation.ComponentScan;
 @EnableActivitiNotificationsGateway
 @EnableActivitiGraphQLQueryService
 @ComponentScan({"org.activiti.cloud.services.query.app","org.activiti.cloud.services.query.events",
-"org.activiti.cloud.services.query.rest","org.activiti.cloud.starters","org.activiti.cloud.starter","org.activiti.cloud.services.identity"})
+"org.activiti.cloud.services.query.rest","org.activiti.cloud.services.security","org.activiti.cloud.starters","org.activiti.cloud.starter","org.activiti.cloud.services.identity"})
 public class QueryApplication {
 
     public static void main(String[] args) {
